TINY-9310: fix test
